+++ chainerx_cc/chainerx/cuda/cuda_device/indexing.cu
@@ -22,6 +22,7 @@
 #include "chainerx/dtype.h"
 #include "chainerx/indexable_array.h"
 #include "chainerx/indexer.h"
+#include "chainerx/kernels/indexing.h"
 #include "chainerx/macro.h"
 #include "chainerx/routines/indexing.h"
 #include "chainerx/shape.h"
